+++ src/core/component/engines/vue3/component.ts
@@ -122,7 +122,7 @@ export function getComponent(meta: ComponentMeta): ComponentOptions<typeof Compo
 		},
 
 		beforeUnmount(): void {
-			init.beforeDestroyState(getComponentContext(this), false);
+			init.beforeDestroyState(getComponentContext(this), {recursive: false});
 		},
 
 		unmounted(): void {