+++ src/core/component/init/states/before-create.ts
@@ -24,7 +24,7 @@ import { implementEventEmitterAPI } from 'core/component/event';
 import { beforeDestroyState } from 'core/component/init/states/before-destroy';
 import { destroyedState } from 'core/component/init/states/destroyed';
 
-import type { ComponentInterface, ComponentMeta, ComponentElement } from 'core/component/interface';
+import type { ComponentInterface, ComponentMeta, ComponentElement, ComponentDestructorOptions } from 'core/component/interface';
 import type { InitBeforeCreateStateOptions } from 'core/component/init/interface';
 
 /**
@@ -65,8 +65,8 @@ export function beforeCreateState(
 		configurable: true,
 		enumerable: false,
 		writable: true,
-		value: (recursive: boolean = true) => {
-			beforeDestroyState(component, recursive);
+		value: (opts: ComponentDestructorOptions) => {
+			beforeDestroyState(component, opts);
 			destroyedState(component);
 		}
 	});